pyproject.toml
[project]
name = "huli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-genai>=1.43.0",
    "langchain[google-genai]>=0.3.27",
    "python-dotenv>=1.1.1",
]
